[RW-7652][risk=no] Mitigate egress suspension timeouts
